trading_ai_assistant/requirements.txt
pyautogui==0.9.54
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
PyQt6==6.6.0
python-dotenv==1.0.0
yfinance==0.2.31  # For fetching market data
ta==0.10.2  # Technical analysis indicators
pyqtgraph==0.13.3
ccxt==4.1.13
pandas-ta==0.3.14b0
rich==13.7.0
questionary==2.0.1
